+++ example/src/utils.ts
@@ -1,9 +1,9 @@
-export function unwrap<T>(thing: T | undefined | null): NonNullable<T> {
-  if (thing === undefined) {
+export function unwrap<T>($: T | undefined | null): NonNullable<T> {
+  if ($ === undefined) {
     throw new Error("Unwrapping failed: value is undefined");
-  } else if (thing === null) {
+  } else if ($ === null) {
     throw new Error("Unwrapping failed: value is null");
   }
 
-  return thing;
+  return $;
 }